Teste de integracao com test containers

+++ api-transaction/src/main/resources/db/migration/V2__create_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS  person (
-    `id` number (20) NOT NULL AUTO_INCREMENT,
+    `id` INT(10) NOT NULL AUTO_INCREMENT,
     `first_name` varchar(80) NOT NULL,
     `last_name` varchar(80) NOT NULL,
     `address` varchar(100) NOT NULL,
